requirements.txt
# Conda packages (not installed via pip, listed here for reference)
# ca-certificates=2023.7.22=h56e8100_0
# cudatoolkit=11.2.2=h933977f_10
# cudnn=8.1.0.77=h3e0f4f4_0
# openssl=3.0.11=h2bbff1b_2
# python=3.9.18=h1aa4202_0
# setuptools=68.0.0=py39haa95532_0
# sqlite=3.41.2=h2bbff1b_0
# tzdata=2023c=h04d1e81_0
# vc=14.2=h21ff451_1
# vs2015_runtime=14.27.29016=h5e58377_2
# wheel=0.41.2=py39haa95532_0

# Pip packages
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
keras==2.10.0
keras-preprocessing==1.1.2
libclang==16.0.6
markdown==3.5.1
markupsafe==2.1.3
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pip==23.3.1
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm~=4.66.1
typing-extensions==4.8.0
urllib3==2.0.7
werkzeug==3.0.1
wrapt==1.15.0
zipp==3.17.0
tensorrt
pandas